Fix a few incorrect tests found with Eastwood lint tool
In query_string_query_test.clj, the same name was used for two
deftests, which causes tests in all but the last deftest with the same
name to be ignored.

update_test.clj had (= x y) expression without an (is ...) around it,
so it could never cause tests to fail.

facets_test.clj had incorrect parentheses leading to two expressions
of the form (is (>= expr)), which can never fail unless an exception
is thrown.