Add scan-and-scroll-seq function.
This takes care of the inconsistency between scroll-seq and ES's
scan-and-scroll API, referenced in issue #72. The new function makes an
initial request of search_type "scan" to the search api in order to
obtain the first scroll id, makes a second request to retrieve the first
batch of results, and then hands off to scroll-seq to continue lazily
retrieving batches. It is necessary to retrieve one batch initially
because the first request is a special case, as described in issue #72
-- it always returns no hits.