Make it possible to request specific nodes stats groups
Before this, the `clojurewerkz.elastisch.rest/nodes-stats-url`
function ignored the `attrs` parameter, making it impossible to pass in
a list of specific nodes stats groups.
There were tests for this scenario but they were passing for the wrong
reason. Subsequent calls to the nodes stats endpoint always return
slightly different responses because of a few variable fields.

Issue #243